+++ src/sdkite/http/__init__.py
@@ -1,4 +1,5 @@
 from sdkite.http.adapter import HTTPAdapter, HTTPAdapterSpec
+from sdkite.http.auth import BasicAuth, NoAuth
 from sdkite.http.model import (
     HTTPBodyEncoding,
     HTTPHeaderDict,
@@ -7,10 +8,12 @@ from sdkite.http.model import (
 )
 
 __all__ = (
+    "BasicAuth",
     "HTTPAdapter",
     "HTTPAdapterSpec",
     "HTTPBodyEncoding",
     "HTTPHeaderDict",
     "HTTPRequest",
     "HTTPResponse",
+    "NoAuth",
 )